Move secret-store test to a different module
